restructure image module some more
